Added exit handler and simplified CLI entrypoint 🥊
