+++ jest.config.ts
@@ -6,8 +6,8 @@ const config: JestConfigWithTsJest = {
   clearMocks: true,
   collectCoverageFrom: [
     'src/**/*',
+    '!src/(index|program).ts',
     // TODO: Write tests for these files when they are less likely to change
-    '!src/index.ts',
     '!src/linters/index.ts',
   ],
   coverageDirectory: 'coverage',
